+++ ontimize/components/table/extensions/header/table-columns-filter/o-table-columns-filter.component.ts
@@ -1,5 +1,5 @@
 import { Component, Inject, forwardRef, OnInit, Injector } from '@angular/core';
-import { OTableComponent } from '../../../o-table.component';
+import { OTableComponent, OColumn } from '../../../o-table.component';
 import { Util } from '../../../../../utils';
 
 export const DEFAULT_INPUTS_O_TABLE_COLUMN_FILTER = [
@@ -22,9 +22,14 @@ export class OTableColumnsFilterComponent implements OnInit {
 
   public static DEFAULT_INPUTS_O_TABLE_COLUMN_FILTER = DEFAULT_INPUTS_O_TABLE_COLUMN_FILTER;
   public static DEFAULT_OUTPUTS_O_TABLE_COLUMN_FILTER = DEFAULT_OUTPUTS_O_TABLE_COLUMN_FILTER;
+  public static COMPARISON_TYPE_SEPARATOR = ':';
+
+  public static DEFAULT_COMPARISON_TYPE = 'VIEW';
+  public static MODEL_COMPARISON_TYPE = 'MODEL';
 
   protected columns: string;
   protected columnsArray: Array<string> = [];
+  protected columnsComparisonProperty: Object = {};
 
   constructor(
     protected injector: Injector,
@@ -34,6 +39,32 @@ export class OTableColumnsFilterComponent implements OnInit {
 
   ngOnInit() {
     this.columnsArray = Util.parseArray(this.columns);
-    this.table.setFilterableColumns(this.columnsArray);
+    if (this.columnsArray.length === 0) {
+      this.columnsArray = this.table.oTableOptions.visibleColumns;
+    }
+    const self = this;
+    this.columnsArray.map((colData, i, arr) => {
+      let colDef = colData.split(OTableColumnsFilterComponent.COMPARISON_TYPE_SEPARATOR);
+      let colName = colDef[0];
+      let compType = (colDef[1] || '').toUpperCase();
+      if ([OTableColumnsFilterComponent.DEFAULT_COMPARISON_TYPE, OTableColumnsFilterComponent.MODEL_COMPARISON_TYPE].indexOf(compType) === -1) {
+        compType = OTableColumnsFilterComponent.DEFAULT_COMPARISON_TYPE;
+      }
+      arr[i] = colName;
+      self.columnsComparisonProperty[colName] = compType;
+    });
+    this.table.setOTableColumnsFilter(this);
+  }
+
+  isColumnFilterable(attr: string) {
+    return (this.columnsArray.indexOf(attr) !== -1);
+  }
+
+  getColumnComparisonValue(column: OColumn, val: any): any {
+    if (!column || this.columnsComparisonProperty[column.attr] === OTableColumnsFilterComponent.MODEL_COMPARISON_TYPE) {
+      return val;
+    } else {
+      return column.renderer ? column.renderer.getCellData(val) : val;
+    }
   }
 }